// src/app/models/productReference.ts
export class productReferenceTS {
 
  id: number;
 
  printcolorname: string;
  tagcolorname: string;
  stickercolorname: string;
  wovencolorname: string;
 
  printreferencenumber: string;
  tagreferencenumber: string;
  stickerreferencenumber: string;
  wovenreferencenumber: string;

  printNum: string;
  tagNum: string;
  wovennum: string;
  stickerNum: string;

  printnext:string;
  tagnext:string;
  stickernext:string;
  wovennext:string;
}